+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - Added `security.separate_fullname` function.
   [sgeulette]
+- Added log parameter in `batching.batch_get_keys`
+  [sgeulette]
 
 1.3.7 (2025-11-05)
 ------------------